+++ vite.config.ts
@@ -25,10 +25,12 @@ export default defineConfig({
     plugins: [],
     test: {
         reporters: process.env.GITHUB_ACTIONS ? [ "dot", "github-actions" ] : [ "dot" ],
+        setupFiles: "./tests/vitest.setup.ts",
         coverage: {
             provider: "v8",
             reporter: [ "text", "json-summary", "json" ],
             include: [ "src/**/*.ts" ],
+            exclude: [ "src/services/api/**", "src/models/**" ],
             reportOnFailure: true,
         },
         alias: [